+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- OfficeOnlineServerHost
+  - Adds new resource OfficeOnlineServerHost
+
 ### Changed
 
 - OfficeOnlineServerDsc